+++ src/modules/auth/jwt.strategy.ts
@@ -8,7 +8,8 @@ import { AuthEntity } from './entities/auth.entity';
 export class JwtStrategy extends PassportStrategy(Strategy) {
   constructor() {
     super({
-      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+      // jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+      jwtFromRequest: ExtractJwt.fromHeader('Authorization'),
       ignoreExpiration: false,
       // secretOrKey: jwtConstants.secret,
       secretOrKey: process.env.JWT_SECRET_KEY,
