--- src/modules/logging/middleware/log-middleware.ts
@@ -1,24 +1,0 @@
-import { Injectable, Logger, NestMiddleware } from '@nestjs/common';
-import { Request, Response, NextFunction } from 'express';
-
-@Injectable()
-export class LogMiddleware implements NestMiddleware {
-  private logger = new Logger('HTTP');
-
-  // use(req: any, res: any, next: () => void) {
-  use(request: Request, response: Response, next: NextFunction): void {
-    const { ip, method, originalUrl } = request;
-    const userAgent = request.get('user-agent') || '';
-
-    response.on('finish', () => {
-      const { statusCode } = response;
-      const contentLength = response.get('content-length');
-
-      this.logger.log(
-        `${method} ${originalUrl} ${statusCode} ${contentLength} - ${userAgent} ${ip}`,
-      );
-    });
-
-    next();
-  }
-}
